@@ -26,22 +26,22 @@ const _defaultLoadingText = "请稍候";
 
 /// show progress dialog with [msg],
 ProgressFuture showProgressDialog({
-  BuildContext context,
-  Widget loading,
-  String loadingText,
-  TextStyle textStyle,
-  Color backgroundColor,
-  double radius,
-  VoidCallback onDismiss,
-  TextDirection textDirection,
-  ProgressOrientation orientation,
+  BuildContext? context,
+  Widget? loading,
+  String? loadingText,
+  TextStyle? textStyle,
+  Color? backgroundColor,
+  double? radius,
+  VoidCallback? onDismiss,
+  TextDirection? textDirection,
+  ProgressOrientation? orientation,
 }) {
   context ??= _contextMap.values.first;
-  _ProgressTheme theme = _ProgressTheme.of(context);
-  theme ??= _ProgressTheme.origin();
+  _ProgressTheme? theme = _ProgressTheme.of(context);
+  theme ??= _ProgressTheme.origin;
   textStyle ??= Theme.of(context)
       .textTheme
-      .bodyText1
+      .bodyText1!
       .copyWith(fontSize: 16.0, color: Colors.white);
   backgroundColor ??= theme.backgroundColor;
   radius ??= theme.radius;
@@ -58,7 +58,7 @@ ProgressFuture showProgressDialog({
           padding: EdgeInsets.symmetric(horizontal: 24.0, vertical: 16.0),
           decoration: BoxDecoration(
               color: backgroundColor,
-              borderRadius: BorderRadius.circular(radius)),
+              borderRadius: BorderRadius.circular(radius!)),
           child: ClipRect(
             child: orientation == ProgressOrientation.vertical
                 ? Column(
@@ -71,7 +71,7 @@ ProgressFuture showProgressDialog({
                         padding: EdgeInsets.all(4.0),
                         child: CircularProgressIndicator(strokeWidth: 3.0),
                       ),
-                      Text(loadingText,
+                      Text(loadingText!,
                           style: textStyle, textAlign: TextAlign.center),
                     ],
                   )
@@ -86,7 +86,7 @@ ProgressFuture showProgressDialog({
                         padding: EdgeInsets.all(4.0),
                         child: CircularProgressIndicator(strokeWidth: 3.0),
                       ),
-                      Text(loadingText,
+                      Text(loadingText!,
                           style: textStyle, textAlign: TextAlign.center),
                     ],
                   ),
@@ -109,7 +109,7 @@ ProgressFuture showProgressDialog({
                           padding: EdgeInsets.all(4.0),
                           child: CupertinoActivityIndicator(),
                         ),
-                        Text(loadingText,
+                        Text(loadingText!,
                             style: textStyle, textAlign: TextAlign.center),
                       ],
                     )
@@ -123,7 +123,7 @@ ProgressFuture showProgressDialog({
                           padding: EdgeInsets.all(4.0),
                           child: CupertinoActivityIndicator(),
                         ),
-                        Text(loadingText,
+                        Text(loadingText!,
                             style: textStyle, textAlign: TextAlign.center),
                       ],
                     ),
@@ -143,17 +143,17 @@ ProgressFuture showProgressDialog({
 /// show [widget] with progress dialog
 ProgressFuture showProgressDialogWidget(
   Widget widget, {
-  BuildContext context,
-  VoidCallback onDismiss,
-  TextDirection textDirection,
-  bool handleTouch,
+  BuildContext? context,
+  VoidCallback? onDismiss,
+  TextDirection? textDirection,
+  bool? handleTouch,
 }) {
   context ??= _contextMap.values.first;
   OverlayEntry entry;
   ProgressFuture future;
 
   var direction = textDirection ??
-      _ProgressTheme.of(context).textDirection ??
+      _ProgressTheme.of(context)!.textDirection ??
       TextDirection.ltr;
 
   GlobalKey<_ProgressContainerState> key = GlobalKey();
@@ -178,7 +178,7 @@ ProgressFuture showProgressDialogWidget(
 
   future = ProgressFuture._(entry, onDismiss, key);
 
-  Overlay.of(context).insert(entry);
+  Overlay.of(context)!.insert(entry);
   ProgressManager().addFuture(future);
   return future;
 }
